codelap/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
